requirements.txt
fastapi==0.75.0
uvicorn==0.17.6
motor==2.5.1
starlette-exporter==0.12.0
pydantic[dotenv]